RADS GUI. Export ASCII GUI. Videos link in XP. GUI improvements  (main map, mouse, etc.). Build scripts improvements and corrections. Other generic changes
RADS
- GUI consolidation.
- control on filter application for huge datasets (RADS, continued)
-  RADS license, disable change output dir when rsync active, etc.etc
- corrections in RADS buttons behaviour.
- solved bug when "sync now" cancelled
- work on refresh rads control panel status
- settings dialog rads control: refresh connection with service
- windows xp rads permissions
- refinements and bug correction in RADS start/stop synchronization

Export ASCII
- consolidation of export ascii GUI.

Videos
- videos for windows xp.

GUI Generic
- correction in application static paths module
- update map on workspace tab changes. (...)
- new template specializations for settings types (gcc picky)
- got rid of v3 settings keywords file, allowing clear unused kewyords in config files.
- mac 3d plot adjustments
- wait mouse improvements
- main map display datsets and filters of selected tab

Generic
- corrections in generic scripts to compile command line tools in CentOS-6
- changed default in scripts for unix systems to Qt 5
- corrections to compile command line tools in CentOS-6
- Improvements in centos6 build script
- corrections in windows build scripts for externals
- adjustments to mac and 32 bit

+++ process/rads/RadsService.cpp
@@ -337,7 +337,7 @@ CRadsClient::CRadsClient( CRadsSettings &settings, QObject *parent )	//parent = 
 	mTimer.start( periodic_check_in_seconds * 1000 );
 
 
-	//startup (starting paused)
+	//startup (starting paused only if PAUSE_ON_START defined for diagnostic)
 
 #if !defined (WAIT_FOR_DEBUGGER) && defined(PAUSE_ON_START)
 
@@ -735,7 +735,7 @@ void CRadsClient::HandleProcessFinished( int exit_code, QProcess::ExitStatus exi
     {
         std::string s = n2s<std::string>( code );
         
-        for ( int i = 0; i <  rsync_exit_codes_size; ++i )
+        for ( size_t i = 0; i <  rsync_exit_codes_size; ++i )
             if ( rsync_exit_codes[i].code == code )
                 return s += ( "-" + rsync_exit_codes[i].msg );
